# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sen Market API - Plateforme d'annonces pour le Sénégal
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all listings with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [date, price]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "401":
          description: Unauthorized
...
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found

    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /listings/user/mine:
    get:
      operationId: getMyListings
      tags: [listings]
      summary: Get current user listings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
        "401":
          description: Unauthorized
  /messages/conversations:
    get:
      operationId: getConversations
      tags: [messages]
      summary: Get all conversations for the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        "401":
          description: Unauthorized
  /messages/conversations/{listingId}/{otherUserId}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get messages for a specific conversation
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
        - name: otherUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
...
  /listings/{id}/renew:
    post:
      operationId: renewListing
      tags: [listings]
      summary: Renew a listing (reset to top of list)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Post a review for a seller
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized
  /reviews/user/{id}:
    get:
      operationId: getUserReviews
      tags: [reviews]
      summary: Get reviews for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - isAdmin
        - createdAt
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        category:
          type: string
        city:
          type: string
        photos:
          type: array
          items:
            type: string
        status:
          type: string
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - city
        - status
        - userId
        - createdAt
        - updatedAt
    ListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        city:
          type: string
        photos:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - category
        - city
    ListingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        city:
          type: string
        photos:
          type: array
          items:
            type: string
        status:
          type: string
    ListingsPage:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - totalPages
...
    Message:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        senderId:
          type: integer
        receiverId:
          type: integer
        listingId:
          type: integer
        sender:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
      required:
        - id
        - content
        - senderId
        - receiverId
        - listingId
        - createdAt
    MessageInput:
      type: object
      properties:
        content:
          type: string
        receiverId:
          type: integer
        listingId:
          type: integer
      required:
        - content
        - receiverId
        - listingId
    Conversation:
      type: object
      properties:
        listingId:
          type: integer
        listing:
          $ref: "#/components/schemas/Listing"
        otherUser:
          $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
      required:
        - listingId
        - otherUser
        - unreadCount
...
    Review:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
        fromUser:
          $ref: "#/components/schemas/User"
      required:
        - id
        - fromUserId
        - toUserId
        - rating
        - createdAt
    ReviewInput:
      type: object
      properties:
        toUserId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
      required:
        - toUserId
        - rating
...