# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sen Market API - Plateforme d'annonces pour le Sénégal
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Review {
  id: number;
  fromUserId: number;
  toUserId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
  fromUser?: User;
}