# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sen Market API - Plateforme d'annonces pour le Sénégal
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  toUserId: number;
  rating: number;
  comment?: string;
}